+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/actioncard/MLBusinessActionCardViewData.kt
@@ -1,4 +1,4 @@
-package com.mercadolibre.android.mlbusinesscomponents.components.splitpayment
+package com.mercadolibre.android.mlbusinesscomponents.components.actioncard
 
 interface MLBusinessActionCardViewData {
     fun getTitle(): String
